refactor: abstract collect_parts for parquet table.
